requirements.txt
# Core dependencies
openai>=1.0.0,<2.0.0  # Pinning major version to avoid breaking changes
streamlit>=1.28.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# LangChain ecosystem (pinned to compatible versions)
langchain>=0.0.340,<0.1.0
langchain-community>=0.0.14,<0.1.0
langchain-core>=0.1.0,<0.2.0

# Additional required dependencies
sqlalchemy>=2.0.0,<3.0.0  # For SQLChatMessageHistory
tiktoken>=0.5.0,<1.0.0  # Used by OpenAI tokenizer